+++ src/error.rs
@@ -1,0 +1,12 @@
+#[derive(Debug)]
+pub enum DhtError<E> {
+    PinError(E),
+    ChecksumMismatch,
+    Timeout,
+}
+
+impl<E> From<E> for DhtError<E> {
+    fn from(error: E) -> DhtError<E> {
+        DhtError::PinError(error)
+    }
+}
