+++ src/pin.rs
@@ -1,0 +1,4 @@
+use embedded_hal::digital::v2::{InputPin, OutputPin};
+
+pub trait InputOutputPin<E>: InputPin<Error = E> + OutputPin<Error = E> {}
+impl<T, E> InputOutputPin<E> for T where T: InputPin<Error = E> + OutputPin<Error = E> {}
